requirements.txt
openai
pydantic
numpy
wmi
pywin32
scikit-learn
pillow
